aurelio e arrumando a possivel falta de permissão do bot

+++ Bot/Comandos/Moderacao.cs
@@ -47,39 +47,45 @@ namespace Bot.Comandos
                 if(!Contexto.IsPrivate){
                     SocketGuildUser author = Contexto.User as SocketGuildUser;
                     if(author.GuildPermissions.ManageMessages){
-                        if(Comando.Length > 1){
-                            Tuple <IUser, string> resUser = null;
-                            if(Comando.Length > 2){
-                                resUser = new Extensions.UserExtensions().GetUser(await Contexto.Guild.GetUsersAsync(), Comando[2]);
-                            }
-                            try{
-                                uint quantidade = Convert.ToUInt32(Comando[1]);
-                                List<IMessage> mensagens = null;
-                                if(resUser != null){
-                                    List<IMessage> construtor = new List<IMessage>();
-                                    IMessage msgRef = Contexto.Message;
-                                    while(construtor.Count < quantidade){
-                                        List<IMessage> temp = (await Contexto.Channel.GetMessagesAsync().FlattenAsync()).ToList();
-                                        construtor.AddRange(temp.FindAll(x => x.Author == resUser.Item1));
-                                    }
-                                    mensagens = construtor;
-                                }else {
-                                    mensagens = (await Contexto.Channel.GetMessagesAsync(limit: Convert.ToInt32(quantidade)).FlattenAsync()).ToList();
+                        if((await Contexto.Guild.GetCurrentUserAsync()).GetPermissions(Contexto.Channel as SocketGuildChannel).ManageMessages) {
+                            if(Comando.Length > 1){
+                                Tuple <IUser, string> resUser = null;
+                                if(Comando.Length > 2){
+                                    resUser = new Extensions.UserExtensions().GetUser(await Contexto.Guild.GetUsersAsync(), Comando[2]);
                                 }
-                                await((ITextChannel)Contexto.Channel).DeleteMessagesAsync(mensagens);
-                            }catch{
-                                await Erro.EnviarErroAsync(await StringCatch.GetStringAsync("limparchatQuantidadeInvalida", "A quantidade de mensagens digitada não é um numero válido"), new DadosErro(await StringCatch.GetStringAsync("limparchatQuantidadeInvalidaArgs", "quantidade usuario"), await StringCatch.GetStringAsync("limparchatQuantidadeInvalidaExemp", "20 @Yummi#1281")));
+                                try{
+                                    uint quantidade = Convert.ToUInt32(Comando[1]);
+                                    List<IMessage> mensagens = null;
+                                    if(resUser != null){
+                                        List<IMessage> construtor = new List<IMessage>();
+                                        IMessage msgRef = Contexto.Message;
+                                        while(construtor.Count < quantidade){
+                                            List<IMessage> temp = (await Contexto.Channel.GetMessagesAsync().FlattenAsync()).ToList();
+                                            construtor.AddRange(temp.FindAll(x => x.Author == resUser.Item1));
+                                        }
+                                        mensagens = construtor;
+                                    }else {
+                                        mensagens = (await Contexto.Channel.GetMessagesAsync(limit: Convert.ToInt32(quantidade) + 1).FlattenAsync()).ToList();
+                                    }
+                                    await((ITextChannel)Contexto.Channel).DeleteMessagesAsync(mensagens);
+                                }catch{
+                                    await Erro.EnviarErroAsync(await StringCatch.GetStringAsync("limparchatQuantidadeInvalida", "a quantidade de mensagens informada não é um numero válido."), new DadosErro(await StringCatch.GetStringAsync("limparchatQuantidadeInvalidaArgs", "quantidade usuario"), await StringCatch.GetStringAsync("limparchatQuantidadeInvalidaExemp", "20 @Yummi#1281")));
+                                }
+                            }
+                            else
+                            {
+                                await Erro.EnviarErroAsync(await StringCatch.GetStringAsync("limparchatQuantidadeInvalida", "você precisa por a quantidade de mensagens que deseja apagar."), new DadosErro(await StringCatch.GetStringAsync("limparchatQuantidadeInvalidaArgs", "quantidade usuario"), await StringCatch.GetStringAsync("limparchatQuantidadeInvalidaExemp", "20 @Yummi#1281")));
                             }
                         }
                         else
                         {
-                            await Erro.EnviarErroAsync(await StringCatch.GetStringAsync("limparchatQuantidadeInvalida", "Você precisa por a quantidade de mensagens que deseja apagar"), new DadosErro(await StringCatch.GetStringAsync("limparchatQuantidadeInvalidaArgs", "quantidade usuario"), await StringCatch.GetStringAsync("limparchatQuantidadeInvalidaExemp", "20 @Yummi#1281")));
+                            await Erro.EnviarErroAsync(await StringCatch.GetStringAsync("limparchatBotPermissao", "eu não tenho permissão para apagar mensagens nesse canal 😔"));
                         }
                     }else{
                         await Erro.EnviarFaltaPermissaoAsync(await StringCatch.GetStringAsync("limparchatPermissao", "Gerenciar Mensagens"));
                     }
                 }else{
-                    await Erro.EnviarErroAsync(await StringCatch.GetStringAsync("limparchatQuantidadeInvalida", "Esse comando não pode ser execultado no privado"));
+                    await Erro.EnviarErroAsync(await StringCatch.GetStringAsync("limparchatQuantidadeInvalida", "esse comando não pode ser executado no privado."));
                 }
             }).Start();
             return Task.CompletedTask;
